fix: upgrade csi-provisioner to v5.0.2 on v1.15.0 chart

+++ pkg/smb/controllerserver.go
@@ -191,8 +191,12 @@ func (d *Driver) DeleteVolume(ctx context.Context, req *csi.DeleteVolumeRequest)
 
 			// archive subdirectory under base-dir. Remove stale archived copy if exists.
 			klog.V(2).Infof("archiving subdirectory %s --> %s", internalVolumePath, archivedInternalVolumePath)
-			if err = os.RemoveAll(archivedInternalVolumePath); err != nil {
-				return nil, status.Errorf(codes.Internal, "failed to delete archived subdirectory %s: %v", archivedInternalVolumePath, err.Error())
+			if d.removeArchivedVolumePath {
+				klog.V(2).Infof("removing archived subdirectory at %v", archivedInternalVolumePath)
+				if err = os.RemoveAll(archivedInternalVolumePath); err != nil {
+					return nil, status.Errorf(codes.Internal, "failed to delete archived subdirectory %s: %v", archivedInternalVolumePath, err.Error())
+				}
+				klog.V(2).Infof("removed archived subdirectory at %v", archivedInternalVolumePath)
 			}
 			if err = os.Rename(internalVolumePath, archivedInternalVolumePath); err != nil {
 				return nil, status.Errorf(codes.Internal, "archive subdirectory(%s, %s) failed with %v", internalVolumePath, archivedInternalVolumePath, err.Error())
